Basic_Evals_and_External_Tracing/pyproject.toml
[project]
name = "basic-evals-and-external-tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langfuse>=3.8.1",
    "openai-agents>=0.4.2",
    "openinference-instrumentation-openai-agents>=1.3.0",
]
